Tidy FE loading of additional details
